Using junit4's parametrized tests in smartstore load tests
