+++ libs/test/SmartStoreTest/src/com/salesforce/androidsdk/store/SmartSqlTest.java
@@ -26,7 +26,7 @@
  */
 package com.salesforce.androidsdk.store;
 
-import android.support.test.filters.MediumTest;
+import android.support.test.filters.SmallTest;
 import android.support.test.runner.AndroidJUnit4;
 
 import com.salesforce.androidsdk.smartstore.store.IndexSpec;
@@ -50,7 +50,7 @@ import org.junit.runner.RunWith;
  * Tests for "smart" sql
  */
 @RunWith(AndroidJUnit4.class)
-@MediumTest
+@SmallTest
 public class SmartSqlTest extends SmartStoreTestCase {
 
 	private static final String BUDGET = "budget";
