+++ libs/test/SmartStoreTest/src/com/salesforce/androidsdk/store/SmartStoreFullTextSearchSpeedTest.java
@@ -26,8 +26,7 @@
  */
 package com.salesforce.androidsdk.store;
 
-import android.support.test.filters.SmallTest;
-import android.support.test.runner.AndroidJUnit4;
+import android.support.test.filters.LargeTest;
 import android.util.Log;
 
 import com.salesforce.androidsdk.smartstore.store.IndexSpec;
@@ -39,16 +38,18 @@ import junit.framework.Assert;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Arrays;
+import java.util.Collection;
 
 /**
  * Tests to compare speed of smartstore full-text-search indices with regular indices
  */
-@RunWith(AndroidJUnit4.class)
-@SmallTest
+@RunWith(Parameterized.class)
+@LargeTest
 public class SmartStoreFullTextSearchSpeedTest extends SmartStoreTestCase {
 
     public static final String TAG = "SmartStoreFTSSpeedTest";
@@ -65,14 +66,19 @@ public class SmartStoreFullTextSearchSpeedTest extends SmartStoreTestCase {
     public static final String ANIMALS_SOUP = "animals";
     public static final String TEXT_COL = "text";
 
-    @Before
-    public void setUp() throws Exception {
-        super.setUp();
-    }
+    @Parameterized.Parameter(0) public String testName;
+    @Parameterized.Parameter(1) public int rowsPerAnimal;
+    @Parameterized.Parameter(2) public int matchingRowsPerAnimal;
 
-    @After
-    public void tearDown() throws Exception {
-        super.tearDown();
+    @Parameterized.Parameters(name = "{0}")
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][]{
+                {"Search1000RowsOneMatch", 40, 1},
+                {"Search1000RowsManyMatches", 40, 40},
+                {"Search10000RowsOneMatch", 400, 1},
+                {"Search10000RowsManyMatches", 400, 400} //,
+                // {"testSearch100000RowsOneMatch", 4000, 1} // Slow - uncomment when collecting performance data
+        });
     }
 
     @Override
@@ -81,31 +87,9 @@ public class SmartStoreFullTextSearchSpeedTest extends SmartStoreTestCase {
     }
 
     @Test
-    public void testSearch1000RowsOneMatch() throws JSONException {
-        trySearch(40, 1);
+    public void test() throws JSONException {
+        trySearch(rowsPerAnimal, matchingRowsPerAnimal);
     }
-
-    @Test
-    public void testSearch1000RowsManyMatches() throws JSONException {
-        trySearch(40, 40);
-    }
-
-    @Test
-    public void testSearch10000RowsOneMatch() throws JSONException {
-        trySearch(400, 1);
-    }
-
-    @Test
-    public void testSearch10000RowsManyMatches() throws JSONException {
-        trySearch(400, 400);
-    }
-
-    /*
-    // Slow - uncomment when collecting performance data
-    public void testSearch100000RowsOneMatch() throws JSONException {
-        trySearch(4000, 1);
-    }
-    */
 
     private void trySearch(int rowsPerAnimal, int matchingRowsPerAnimal) throws JSONException {
         double totalInsertTimeString = setupData(Type.string, rowsPerAnimal, matchingRowsPerAnimal);
