+++ libs/test/SmartStoreTest/src/com/salesforce/androidsdk/store/SmartStoreLoadTest.java
@@ -26,143 +26,60 @@
  */
 package com.salesforce.androidsdk.store;
 
-import android.support.test.filters.SmallTest;
-import android.support.test.runner.AndroidJUnit4;
+import android.support.test.filters.LargeTest;
 import android.util.Log;
 
-import com.salesforce.androidsdk.smartstore.store.IndexSpec;
 import com.salesforce.androidsdk.smartstore.store.QuerySpec;
-import com.salesforce.androidsdk.smartstore.store.SmartStore;
 import com.salesforce.androidsdk.smartstore.store.SmartStore.Type;
 
 import org.json.JSONArray;
 import org.json.JSONException;
-import org.json.JSONObject;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 
 /**
  * Set of tests for the smart store loading numerous and/or large entries and querying them back
  */
-@RunWith(AndroidJUnit4.class)
-@SmallTest
-public class SmartStoreLoadTest extends SmartStoreTestCase {
+@RunWith(Parameterized.class)
+@LargeTest
+public class SmartStoreLoadTest extends SmartStoreLoadTestCase {
 
-    protected static final String TEST_SOUP = "test_soup";
+    @Parameterized.Parameter(0) public String testName;
+    @Parameterized.Parameter(1) public Type indexType;
+    @Parameterized.Parameter(2) public int numberEntries;
+    @Parameterized.Parameter(3) public int numberFieldsPerEntry;
+    @Parameterized.Parameter(4) public int numberCharactersPerField;
+    @Parameterized.Parameter(5) public int numberIndexes;
 
-    private static final int NUMBER_ENTRIES = 1000;
-    private static final int NUMBER_ENTRIES_PER_BATCH = 100;
-    private static final int NS_IN_MS = 1000000;
-
-    @Before
-    public void setUp() throws Exception {
-        super.setUp();
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        super.tearDown();
-    }
-
-    @Test
-    public void testUpsertQuery1StringIndex1field20characters() throws JSONException {
-        tryUpsertQuery(Type.string, NUMBER_ENTRIES, 1, 20, 1);
-    }
-
-    @Test
-    public void testUpsertQuery1StringIndex1field1000characters() throws JSONException {
-        tryUpsertQuery(Type.string, NUMBER_ENTRIES, 1, 1000, 1);
+    @Parameterized.Parameters(name = "{0}")
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][]{
+                {"UpsertQuery1StringIndex1field20characters", Type.string, NUMBER_ENTRIES, 1, 20, 1},
+                {"UpsertQuery1StringIndex1field1000characters", Type.string, NUMBER_ENTRIES, 1, 1000, 1},
+                {"UpsertQuery1StringIndex10fields20characters", Type.string, NUMBER_ENTRIES, 10, 20, 1},
+                {"UpsertQuery10StringIndexes10fields20characters", Type.string, NUMBER_ENTRIES, 10, 20, 10},
+                {"UpsertQuery1JSON1Index1field20characters", Type.json1, NUMBER_ENTRIES, 1, 20, 1},
+                {"UpsertQuery1JSON1Index1field1000characters", Type.json1, NUMBER_ENTRIES, 1, 1000, 1},
+                {"UpsertQuery1JSON1Index10fields20characters", Type.json1, NUMBER_ENTRIES, 10, 20, 1},
+                {"UpsertQuery10JSON1Indexes10fields20characters", Type.json1, NUMBER_ENTRIES, 10, 20, 10}
+        });
     }
 
     @Test
-    public void testUpsertQuery1StringIndex10fields20characters() throws JSONException {
-        tryUpsertQuery(Type.string, NUMBER_ENTRIES, 10, 20, 1);
-    }
-
-    @Test
-    public void testUpsertQuery10StringIndexes10fields20characters() throws JSONException {
-        tryUpsertQuery(Type.string, NUMBER_ENTRIES, 10, 20, 10);
-    }
-
-    @Test
-    public void testUpsertQuery1JSON1Index1field20characters() throws JSONException {
-        tryUpsertQuery(Type.json1, NUMBER_ENTRIES, 1, 20, 1);
-    }
-
-    @Test
-    public void testUpsertQuery1JSON1Index1field1000characters() throws JSONException {
-        tryUpsertQuery(Type.json1, NUMBER_ENTRIES, 1, 1000, 1);
-    }
-
-    @Test
-    public void testUpsertQuery1JSON1Index10fields20characters() throws JSONException {
-        tryUpsertQuery(Type.json1, NUMBER_ENTRIES, 10, 20, 1);
-    }
-
-    @Test
-    public void testUpsertQuery10JSON1Indexes10fields20characters() throws JSONException {
-        tryUpsertQuery(Type.json1, NUMBER_ENTRIES, 10, 20, 10);
-    }
-
-    @Test
-    public void testAlterSoupClassicIndexing() throws JSONException {
-        tryAlterSoup(Type.string);
-    }
-
-    @Test
-    public void testAlterSoupJSON1Indexing() throws JSONException {
-        tryAlterSoup(Type.json1);
-    }
-
-    protected String getEncryptionKey() {
-        return "";
-    }
-
-    protected String getTag() {
-        return getClass().getSimpleName();
+    public void test() throws JSONException {
+        tryUpsertQuery(indexType, numberEntries, numberFieldsPerEntry, numberCharactersPerField, numberIndexes);
     }
 
     private void tryUpsertQuery(Type indexType, int numberEntries, int numberFieldsPerEntry, int numberCharactersPerField, int numberIndexes) throws JSONException {
         setupSoup(TEST_SOUP, numberIndexes, indexType);
         upsertEntries(numberEntries / NUMBER_ENTRIES_PER_BATCH, NUMBER_ENTRIES_PER_BATCH, numberFieldsPerEntry, numberCharactersPerField);
         queryEntries();
-    }
-
-    protected void setupSoup(String soupName, int numberIndexes, Type indexType) {
-        IndexSpec[] indexSpecs = new IndexSpec[numberIndexes];
-        for (int indexNumber=0; indexNumber<numberIndexes; indexNumber++) {
-            indexSpecs[indexNumber] = new IndexSpec("k_" + indexNumber, indexType);
-        }
-        registerSoup(store, soupName, indexSpecs);
-        Log.i(getTag(), String.format("Creating table with %d %s indexes", numberIndexes, indexType));
-    }
-
-    private void upsertEntries(int numberBatches, int numberEntriesPerBatch, int numberFieldsPerEntry, int numberCharactersPerField) throws JSONException {
-        List<Long> times = new ArrayList<Long>();
-        for (int batchNumber=0; batchNumber<numberBatches; batchNumber++) {
-            long start = System.nanoTime();
-            store.beginTransaction();
-            for (int entryNumber=0; entryNumber<numberEntriesPerBatch; entryNumber++) {
-                JSONObject entry = new JSONObject();
-                for (int fieldNumber=0; fieldNumber<numberFieldsPerEntry; fieldNumber++) {
-                    String value = pad( "v_" + batchNumber + "_" + entryNumber + "_" + fieldNumber + "_", numberCharactersPerField);
-                    entry.put("k_" + fieldNumber, value);
-                }
-                store.upsert(TEST_SOUP, entry, SmartStore.SOUP_ENTRY_ID, false);
-            }
-            store.setTransactionSuccessful();
-            store.endTransaction();
-            long end = System.nanoTime();
-            times.add(end - start);
-        }
-        double avgMilliseconds = average(times) / NS_IN_MS;
-        Log.i(getTag(), String.format("Upserting %d entries with %d per batch with %d fields with %d characters: average time per batch --> %.3f ms",
-                numberBatches * numberEntriesPerBatch, numberEntriesPerBatch, numberFieldsPerEntry, numberCharactersPerField, avgMilliseconds));
     }
 
     private void queryEntries() throws JSONException {
@@ -200,48 +117,5 @@ public class SmartStoreLoadTest extends SmartStoreTestCase {
         double avgMilliseconds = average(times) / NS_IN_MS;
         Log.i(getTag(), String.format("Querying with %s query matching %d entries and %d page size: average time per page --> %.3f ms",
                 querySpec.queryType, countMatches, querySpec.pageSize, avgMilliseconds));
-    }
-
-    private String pad(String s, int numberCharacters) {
-        StringBuffer sb = new StringBuffer(numberCharacters);
-        sb.append(s);
-        for (int i=s.length(); i<numberCharacters; i++) {
-            sb.append("x");
-        }
-        return sb.toString();
-    }
-
-    private double average(List<Long> times) {
-        double avg = 0;
-        for (int i=0; i<times.size(); i++) {
-            avg += times.get(i);
-        }
-        avg /= times.size();
-        return avg;
-    }
-
-    private void tryAlterSoup(Type indexType) throws JSONException {
-        Log.i(getTag(), "In testAlterSoup");
-        Log.i(getTag(), String.format("Initial database size: %d bytes", store.getDatabaseSize()));
-        setupSoup(TEST_SOUP, 1, indexType);
-        upsertEntries(NUMBER_ENTRIES / NUMBER_ENTRIES_PER_BATCH, NUMBER_ENTRIES_PER_BATCH, 10, 20);
-        Log.i(getTag(), String.format("Database size after: %d bytes", store.getDatabaseSize()));
-
-        // Without indexing for new index specs
-        alterSoup("Adding one index / no re-indexing", false, new IndexSpec[]{new IndexSpec("k_0", indexType), new IndexSpec("k_1", indexType)});
-        alterSoup("Adding one index / dropping one index / no re-indexing", false, new IndexSpec[] {new IndexSpec("k_0", indexType), new IndexSpec("k_2", indexType)});
-        alterSoup("Dropping one index / no re-indexing", false, new IndexSpec[] {new IndexSpec("k_0", indexType)});
-
-        // With indexing for new index specs
-        alterSoup("Adding one index / with re-indexing", true, new IndexSpec[] {new IndexSpec("k_0", indexType), new IndexSpec("k_1", indexType)});
-        alterSoup("Adding one index / dropping one index / with re-indexing", true, new IndexSpec[] {new IndexSpec("k_0", indexType), new IndexSpec("k_2", indexType)});
-        alterSoup("Dropping one index / with re-indexing", true, new IndexSpec[] {new IndexSpec("k_0", indexType)});
-    }
-    
-    private void alterSoup(String msg, boolean reIndexData, IndexSpec[] indexSpecs) throws JSONException {
-        long start = System.nanoTime();
-        store.alterSoup(TEST_SOUP, indexSpecs, reIndexData);
-        double duration = System.nanoTime() - start;
-        Log.i(getTag(), String.format("%s completed in: %.3f ms", msg, duration/ NS_IN_MS));
     }
 }
